backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
fireworks-ai
sentence-transformers
faiss-cpu
pypdf
python-dotenv